+++ lib/typed-object-entries.ts
